Coomit of my linked list implementation
The file "linkedList.h" contains the definitions and implementation of the linked list data structure. 
The file "linkedList.cpp" contains the program to use those functions.
The file "linkedList" is the output file of the "linkedList.cpp".
The file "detectCycle.cpp" is the file used to find out if a cycle exists or not in the given linked list.
The file "detectCycle" is the output file of the "detectCycle.cpp".